Move storage.Service to own file
