+++ src/storage/index.ts
@@ -2,12 +2,4 @@
  * @module "ethpm/storage"
  */
 
-import { Maybe } from "types";
-import { URL } from "url";
-
-export interface Service {
-  read (uri: URL): Promise<Maybe<string>>;
-  hash (content: string): Promise<string>;
-  predictUri (content: string): Promise<URL>;
-  write (content: string): Promise<URL>;
-}
+export * from "./service";
