+++ src/storage/service.ts
@@ -1,0 +1,13 @@
+/**
+ * @module "ethpm/storage"
+ */
+
+import { Maybe } from "types";
+import { URL } from "url";
+
+export interface Service {
+  read (uri: URL): Promise<Maybe<string>>;
+  hash (content: string): Promise<string>;
+  predictUri (content: string): Promise<URL>;
+  write (content: string): Promise<URL>;
+}
